+++ R/update.R
@@ -5,12 +5,13 @@
 #' Documents are identified by the \code{query} or
 #' by `_id`'s in \code{value}, where the latter takes
 #' precedence.
-#' \code{value} can have multiple documents and `_id`'s,
-#' which then are used for iterative updating.
+#' \code{value} can have multiple documents and
+#' `_id`'s, which then are used for iterative updating.
 #'
-#' Uses native functions with MongoDB, SQLite, DuckDB
-#' and Elasticsearch; a `plpgsql` function added to
-#' PostgreSQL; and [jqr] for CouchDB.
+#' Uses native functions with MongoDB, SQLite, and
+#' DuckDB; delete and create with Elasticsearch;
+#' a `plpgsql` function added to PostgreSQL;
+#' and [jqr] for CouchDB.
 #'
 #' @inheritParams docdb_create
 #'
@@ -45,7 +46,9 @@ docdb_update <- function(src, key, value, query, ...) {
   assert(src, "docdb_src")
   assert(key, "character")
   assert(value, c("data.frame", "list", "character"))
-  assert(query, "character")
+  assert(query, c("json", "character"))
+  stopifnot(jsonlite::validate(query))
+
   UseMethod("docdb_update", src)
 }
 
@@ -59,44 +62,129 @@ docdb_update.src_couchdb <- function(src, key, value, query, ...) {
   # and save the entire new revision (or version) of that document back into CouchDB.
 
   # get original set
-  input <- docdb_query(src, key, query)
+  query <- jsonlite::minify(query)
 
-  # early return if not found
-  if (!length(input)) return(0L)
+  # process value, target is json strings in file
 
-  # original set data frame to json
-  ndjson <- NULL
-  jsonlite::stream_out(input, con = textConnection(
-    object = "ndjson", open = "w", local = TRUE), verbose = FALSE)
+  # check other inputs
+  if (isFile(value)) {
+    value <- jqr::jq(file(value), ".", flags = jqr::jq_flags(pretty = FALSE))
+  } else if (isUrl(value)) {
+    value <- jqr::jq(url(value), ".", flags = jqr::jq_flags(pretty = FALSE))
+  }
+
+  # handle potential json string input
+  if (length(value) == 1 && is.atomic(value) &&
+      is.character(value) && jsonlite::validate(value)
+  ) {
+    # check format
+    if (all(jqr::jq(value, " .[] | type ") == '"array"') && length(jqr::jq(value, " .[] ")) > 1L) stop(
+      "Require JSON string that is an array of documents, not a set of fields that are arrays."
+    )
+    # check if top level is an array
+    chk <- jqr::jq(value, " type ")
+    if (length(chk) == 1L && chk == '"array"') value <- jqr::jq(value, " .[] ")
+  }
 
   # data frame to json
   if (all(class(value) %in% "data.frame")) {
+    # if value contains id's, split rows into documents of a vector
     row.names(value) <- NULL
-    value <- jsonlite::toJSON(value, dataframe = "columns", auto_unbox = TRUE)
+    if (any(names(value) == "_id")) {
+      value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE, digits = NA)
+      value <- jqr::jq(value, " .[] ")
+    } else {
+      # otherwise keep as single document
+      value <- jsonlite::toJSON(value, dataframe = "columns", auto_unbox = TRUE, digits = NA)
+    }
   }
   # list to json
   if (all(class(value) %in% "list")) {
-    value <- jsonlite::toJSON(value, auto_unbox = TRUE)
+    value <- jsonlite::toJSON(value, auto_unbox = TRUE, digits = NA)
+    # check if {"_id":["Valiant","Fiat 128"],"gear":[8,9]}
+    if (all(jqr::jq(value, '._id | type' ) == '"array"')) {
+      value <- data.frame(jsonlite::fromJSON(value), check.names = FALSE)
+      value <- jsonlite::toJSON(value)
+    }
+    # check if top level is an array
+    chk <- jqr::jq(value, " type ")
+    if (length(chk) == 1L && chk == '"array"') value <- jqr::jq(value, " .[] ")
   }
 
-  # merge json with value
-  value <- jqr::jq(paste0(
-    "[",  jqr::jq(textConnection(ndjson), "."),
-    ",", value, "]"), ' reduce .[] as $item ({}; . * $item) ')
+  # get _id's of docs to be updated
+  valueIds <- gsub(
+    '"', "", as.character(jqr::jq(textConnection(value), '._id | select(. != null)'
+    )))
 
-  # jqr output to list
-  value <- jsonlite::stream_in(
-    textConnection(value),
-    simplifyVector = FALSE,
-    verbose = FALSE)
+  # early return if none found
+  if (query == "{}" && !length(valueIds)) return(0L)
 
-  # note: sofa::db_bulk_update changes all
-  # documents in the container, therefore
+  # check
+  if (query != "{}" && length(valueIds)) warning(
+    "Ignoring the specified 'query' parameter, using _id's ",
+    "found in 'value' to identify documents to be updated")
+
+  # get docs to be updated
+  if (length(valueIds)) query <- paste0(
+    '{"_id": {"$in": [', paste0('"', valueIds, '"', collapse = ", "), ']}}')
+  #
+  input <- docdb_query(src, key, query)
+  if (!length(input)) return(0L)
+  if (!length(valueIds)) valueIds <- input[["_id"]]
+  #
+  ndjson <- NULL
+  jsonlite::stream_out(input, con = textConnection(
+    object = "ndjson", open = "w", local = TRUE), verbose = FALSE, digits = NA)
+
+  # temporary file and connection
+  tfname <- tempfile()
+  tfnameCon <- file(tfname, open = "at")
+  # register to close and remove file after used for streaming
+  on.exit(unlink(tfname), add = TRUE)
+  on.exit(try(close(tfnameCon), silent = TRUE), add = TRUE)
+
+  # check
+  if (length(value) > 1L && !identical(nrow(input), length(value))) stop(
+    "Unequal number of documents identified (", nrow(input),
+    ") and of documents in 'value' (", length(value), ")"
+  )
+
+  # merge and append to file
+  for (i in seq_along(valueIds)) {
+
+    idValue <- jqr::jq(
+      textConnection(value),
+      paste0('select(._id == "', valueIds[i], '")'))
+
+    if (!length(idValue)) {
+      if (length(value) > 1L) {
+        idValue <- value[i]
+      } else {
+        idValue <- value
+      }
+    }
+
+    # merge input json with value
+    jqr::jq(
+      x = textConnection(paste0(
+        "[", jqr::jq(
+          textConnection(ndjson),
+          paste0("select(._id == \"", valueIds[i], "\")")),
+        ",", idValue, "]")),
+      " reduce .[] as $item ({}; . * $item) ",
+      out = tfnameCon)
+
+  }
+
+  close(tfnameCon)
+
+  # note: sofa::db_bulk_update would change all
+  # documents in the container, therefore do:
   # - delete documents
   invisible(docdb_delete(src, key, query = query, ...))
   # - create documents
   result <- suppressMessages(
-    docdb_create(src, key, value, ...))
+    docdb_create(src, key, tfname, ...))
 
   # return
   return(result)
@@ -106,29 +194,63 @@ docdb_update.src_couchdb <- function(src, key, value, query, ...) {
 #' @export
 docdb_update.src_elastic <- function(src, key, value, query, ...) {
 
-  # get _id's
-  ids <- docdb_query(src, key, query, fields = '{"_id": 1}')[["_id"]]
-
-  # early return if not found
-  if (!length(ids)) return(0L)
-
   # process value, target is data frame
+  query <- jsonlite::minify(query)
+
+  # needed because target is data frame
+  # but data frame could be serialised
+  # into a single json string line, in
+  # contrast to other inputs
+  valueClass <- class(value)
 
   # json to data frame
-  if (all(class(value) %in% "character")) {
-    value <- jsonlite::fromJSON(value)
+  if (any(class(value) == "character")) {
+    if (isFile(value)) {
+      value <- jsonlite::stream_in(file(value), verbose = FALSE)
+    } else {
+      if (isUrl(value)) {
+        value <- jsonlite::stream_in(url(value), verbose = FALSE)
+      } else {
+        value <- jsonlite::stream_in(textConnection(value), verbose = FALSE)
+      }
+    }
   }
 
   # list to data frame
   if (all(class(value) %in% "list")) {
     value <- jsonlite::fromJSON(
-      jsonlite::toJSON(value, auto_unbox = TRUE))
+      jsonlite::toJSON(value, auto_unbox = TRUE, digits = NA))
     # if value is still simple list
-    value <- as.data.frame(value)
+    value <- as.data.frame(value, check.names = FALSE)
   }
 
   # data frame
   row.names(value) <- NULL
+  if (any(names(value) == "X_id")) names(value)[names(value) == "X_id"] <- "_id"
+
+  # _id in data frame?
+  if (any(names(value) == "_id")) {
+
+    ids <- value[, "_id", drop = TRUE]
+    value <- value[, -match("_id", names(value)), drop = FALSE]
+
+    if (query != '{}') warning(
+      "Ignoring the specified 'query' parameter, using _id's ",
+      "found in 'value' to identify documents to be updated")
+
+  } else {
+
+    ids <- docdb_query(src, key, query, fields = '{"_id": 1}')[["_id"]]
+    if (!length(ids)) return(0L)
+
+  }
+
+  # check
+  if (!all(valueClass %in% "data.frame") &&
+      nrow(value) > 1L && !identical(nrow(value), length(ids))) stop(
+    "Unequal number of documents identified (", length(ids),
+    ") and of documents in 'value' (", nrow(value), ")"
+  )
 
   # how to handle?
   if (nrow(value) != length(ids)) {
@@ -139,7 +261,8 @@ docdb_update.src_elastic <- function(src, key, value, query, ...) {
     if (nrow(indf) == 1L) {
       for (i in seq_len(length(ids))) {
         value <- rbind(value, indf)
-      }}
+      }
+    }
     #
     if (nrow(indf) > 1L) {
       # iterating over _id's
@@ -159,6 +282,7 @@ docdb_update.src_elastic <- function(src, key, value, query, ...) {
     doc_ids = ids,
     quiet = TRUE,
     digits = NA,
+    query = list(refresh = TRUE),
     ...
   )
 
@@ -172,11 +296,12 @@ docdb_update.src_elastic <- function(src, key, value, query, ...) {
   }
   result <- unlist(result, use.names = TRUE)
   result <- result[names(result) == "items.update.result"]
-  if (any(result != "updated")) {
+  if (any(result != "updated" & result != "noop")) {
     warning("Could not create some documents, reason: ",
-            unique(result[result != "updated"]), call. = FALSE, immediate. = TRUE)
+            unique(result[result != "updated" & result != "noop"]),
+            call. = FALSE, immediate. = TRUE)
   }
-  result <- sum(result == "updated")
+  result <- sum(result == "updated" | result == "noop")
   return(result)
 
 }
@@ -186,10 +311,11 @@ docdb_update.src_mongo <- function(src, key, value, query, ...) {
 
   # special check for mongo
   chkSrcMongo(src, key)
+  query <- jsonlite::minify(query)
 
   # if regexp query lacks options, add them in
   if (grepl('"[$]regex" *: *"[^,$:}]+?" *}', query)) query <-
-      sub('("[$]regex" *: *"[^,$:}]+?" *)', '\\1, "$options": ""', query)
+    sub('("[$]regex" *: *"[^,$:}]+?" *)', '\\1, "$options": ""', query)
 
   # process value, target is json string
 
@@ -203,9 +329,9 @@ docdb_update.src_mongo <- function(src, key, value, query, ...) {
   # handle potential json string input
   if (length(value) == 1 && is.atomic(value) &&
       is.character(value) && jsonlite::validate(value)
-      ) {
+  ) {
     # check format
-    if (all(jqr::jq(value, " .[] | type ") == '"array"') & length(jqr::jq(value, " .[] ")) > 1L) stop(
+    if (all(jqr::jq(value, " .[] | type ") == '"array"') && length(jqr::jq(value, " .[] ")) > 1L) stop(
       "Require JSON string that is an array of documents, not a set of fields that are arrays."
     )
     # check if top level is an array
@@ -218,17 +344,17 @@ docdb_update.src_mongo <- function(src, key, value, query, ...) {
     # if value contains id's, split rows into documents of a vector
     row.names(value) <- NULL
     if (any(names(value) == "_id")) {
-      value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE)
+      value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE, digits = NA)
       value <- jqr::jq(value, " .[] ")
     } else {
       # otherwise keep as single document
-      value <- jsonlite::toJSON(value, dataframe = "columns", auto_unbox = TRUE)
+      value <- jsonlite::toJSON(value, dataframe = "columns", auto_unbox = TRUE, digits = NA)
     }
   }
 
   # list to json
   if (all(class(value) %in% "list")) {
-    value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE)
+    value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE, digits = NA)
     # check if top level is an array
     chk <- jqr::jq(value, " type ")
     if (length(chk) == 1L && chk == '"array"') value <- jqr::jq(value, " .[] ")
@@ -241,7 +367,7 @@ docdb_update.src_mongo <- function(src, key, value, query, ...) {
   ids <- gsub("\"", "", as.character(ids))
   ids <- ids[ids != "null"]
   if (length(ids)) {
-    if (query != "") warning(
+    if (query != "{}") warning(
       "Ignoring the specified 'query' parameter, using _id's ",
       "found in 'value' to identify documents to be updated")
     value <- jqr::jq(value, " del(._id) ")
@@ -259,7 +385,7 @@ docdb_update.src_mongo <- function(src, key, value, query, ...) {
 
   # check
   if (length(value) > 1L && (length(value) != length(ids))) stop(
-    "Unequal number of documents identified (", (length(ids)),
+    "Unequal number of documents identified (", length(ids),
     ") and of documents in 'value' (", length(value), ")"
   )
 
@@ -309,6 +435,8 @@ docdb_update.src_mongo <- function(src, key, value, query, ...) {
 #' @export
 docdb_update.src_sqlite <- function(src, key, value, query, ...) {
 
+  query <- jsonlite::minify(query)
+
   # SQL for patching, see https://www.sqlite.org/json1.html#jpatch
   updFunction <- "json_patch"
 
@@ -318,6 +446,8 @@ docdb_update.src_sqlite <- function(src, key, value, query, ...) {
 
 #' @export
 docdb_update.src_postgres <- function(src, key, value, query, ...) {
+
+  query <- jsonlite::minify(query)
 
   # Since PostgreSQL has no internal function,
   # uses function inserted by nodbi::src_postgres
@@ -330,8 +460,10 @@ docdb_update.src_postgres <- function(src, key, value, query, ...) {
 #' @export
 docdb_update.src_duckdb <- function(src, key, value, query, ...) {
 
+  query <- jsonlite::minify(query)
+
   # use file based approach
-  if (isFile(value) && (query == "" | query == "{}")) {
+  if (isFile(value) && (query == "{}")) {
 
     statement <- paste0(
       'UPDATE "', key, '"
@@ -376,7 +508,7 @@ sqlUpdate <- function(src, key, value, query, updFunction) {
   if (length(value) == 1 && is.atomic(value) &&
       is.character(value) && jsonlite::validate(value)) {
     # check format
-    if (all(jqr::jq(value, " .[] | type ") == '"array"') & length(jqr::jq(value, " .[] ")) > 1L) stop(
+    if (all(jqr::jq(value, " .[] | type ") == '"array"') && length(jqr::jq(value, " .[] ")) > 1L) stop(
       "Require JSON string that is an array of documents, not a set of fields that are arrays."
     )
     # check if top level is an array
@@ -389,17 +521,17 @@ sqlUpdate <- function(src, key, value, query, updFunction) {
     # if value contains id's, split rows into documents of a vector
     row.names(value) <- NULL
     if (any(names(value) == "_id")) {
-      value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE)
+      value <- jsonlite::toJSON(value, dataframe = "rows", auto_unbox = TRUE, digits = NA)
       value <- jqr::jq(value, " .[] ")
     } else {
       # otherwise keep as single document
-      value <- jsonlite::toJSON(value, dataframe = "columns", auto_unbox = TRUE)
+      value <- jsonlite::toJSON(value, dataframe = "columns", auto_unbox = TRUE, digits = NA)
     }
   }
 
   # list to json
   if (all(class(value) %in% "list")) {
-    value <- jsonlite::toJSON(value, auto_unbox = TRUE)
+    value <- jsonlite::toJSON(value, auto_unbox = TRUE, digits = NA)
     # check if top level is an array
     chk <- jqr::jq(value, " type ")
     if (length(chk) == 1L && chk == '"array"') value <- jqr::jq(value, " .[] ")
@@ -412,7 +544,7 @@ sqlUpdate <- function(src, key, value, query, updFunction) {
   ids <- gsub("\"", "", as.character(ids))
   ids <- ids[ids != "null"]
   if (length(ids)) {
-    if (query != "" & query != "{}") warning(
+    if (query != "{}") warning(
       "Ignoring the specified 'query' parameter, using _id's ",
       "found in 'value' to identify documents to be updated")
     value <- jqr::jq(value, " del(._id) ")
@@ -473,5 +605,3 @@ sqlUpdate <- function(src, key, value, query, updFunction) {
   return(result)
 
 }
-
-