+++ tests/testthat/helper.R
@@ -1,9 +1,6 @@
 
 #### params for tests ####
 
-# delay needed
-elasticSleep <- 1L # seconds
-
 # Warning (core-nodbi.R:176:3): docdb_update, docdb_query
 # Changing language has no effect when envvar LANG='C'
 Sys.unsetenv("LANG")